// cats-app-front/src/app/interfaces/filter.interface.ts
export interface Filter {
    name?: string;
    origin?: string;
    averageLifespanFrom?: number;
    averageLifespanTo?: number;
    lengthFrom?: number;
    lengthTo?: number;
    weightFrom?: number;
    weightTo?: number;
    gentlenessFrom?: number;
    gentlenessTo?: number;
    immunityFrom?: number;
    immunityTo?: number;
    playfulnessFrom?: number;
    playfulnessTo?: number;
    moltFrom?: number;
    moltTo?: number;
    careFrom?: number;
    careTo?: number;
    intelligenceFrom?: number;
    intelligenceTo?: number;
    childFriendlinessFrom?: number;
    childFriendlinessTo?: number;
    petFriendlinessFrom?: number;
    petFriendlinessTo?: number;
}